backend/generate/requirements.txt
psycopg2-binary>=2.9.0
openai>=1.0.0
boto3>=1.26.0
requests>=2.28.0
